Listener wyciągnięty z CommandManagerWindow i inne poprawki
